add Number of Time Steps to Time Step Control
Allows user to specify number of time steps to take. Along with the
Initial Time, Final Time, the step size is calculated.

Many of the Drekar examples checked only states the Number of Time Steps
that should be taken and defers to the Integrator (Rythmos) to compute
the step size. This commit allows Drekar users to run those same
examples again without too much change.